+++ CHANGELOG.md
@@ -12,6 +12,16 @@
 
 - Add `org.jetbrains.intellij.platform.verifyPluginDefaultRecommendedIdes` Gradle property to control automatic default IDE selection for the `verifyPlugin` task.
 - Add `intellijPlatform.pluginVerification.ides.current()` helper to refer to the currently targeted IntelliJ Platform.
+- Add default value for `targetRootOutputDir` of `GenerateParserTask`.
+- Add properties `targetRootOutputDir` and `packageName` (with defaults) for `GenerateLexerTask`. These properties replace `targetOutputDir` and automatically use a proper subdirectory matching the package.
+
+### Changed
+
+- Deprecate property `targetOutputDir` of `GenerateLexerTask`.
+- Deprecate properties `pathToParser` and `pathToPsiRoot` of `GenerateParserTask`.
+- Purge stale files `GenerateLexerTask` and `GenerateParserTask` by default (as soon as you stop using the deprecated properties).
+- Deprecate `targetFile(String)` and `targetFile(Provider)` of `GenerateLexerTask`.
+- Deprecate `parserFile()` and `psiDir()` of `GenerateParserTask`.
 
 ### Fixed